requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
drf-standardized-errors==0.13.0
drf-yasg==1.21.7
drf-yasg==1.21.7
